+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.1.0] - 2020-10-23
+
+### Added
+
+- Added Text.with_indentation_guide
+- Added Text.detect_indentation
+- Added Pretty.indent_guides
+- Added Syntax.indent_guides
+- Added indent_guides parameter on pretty.install
+- Added rich.pretty.pprint
+- Added max_length to Pretty
+
+### Changed
+
+- Enabled indent guides on Tracebacks
+
 ## [9.0.1] - 2020-10-19
 
 ### Fixed